+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,18 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported Container access policy with new permission x,t
+    -  `New-AzStorageContainerStoredAccessPolicy`
+    -  `Set-AzStorageContainerStoredAccessPolicy`
+* Change the output of get/set Container/Share/Queue/Table access policy cmdlet, by change the child property Permission type from enum to String
+    -  `Get-AzStorageContainerStoredAccessPolicy`
+    -  `Set-AzStorageContainerStoredAccessPolicy`
+    -  `Get-AzStorageShareStoredAccessPolicy`
+    -  `Set-AzStorageShareStoredAccessPolicy`
+    -  `Get-AzStorageQueueStoredAccessPolicy`
+    -  `Set-AzStorageQueueStoredAccessPolicy`
+    -  `Get-AzStorageTableStoredAccessPolicy`
+    -  `Set-AzStorageTableStoredAccessPolicy`
 * Support set/update/remove Acl recursively on Datalake Gen2 item 
     -  `Set-AzDataLakeGen2AclRecursive` 
     -  `Update-AzDataLakeGen2AclRecursive`